requirements.txt
# Zerodha KiteConnect
kiteconnect>=5.0.1

# Data & Numerics
pandas>=2.1.0
numpy>=1.26.0

# HTTP / Async
requests>=2.31.0
urllib3>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.4

# Timezone
pytz>=2024.1

# CLI / Output
tabulate>=0.9.0
colorama>=0.4.6
rich>=13.7.0

# Plotting / Reports
matplotlib>=3.8.0
seaborn>=0.13.0

# Data storage
pyarrow>=14.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
